Use OpenAPI client generation for unit tests.
