+++ dotnet-monitor.sln
@@ -29,6 +29,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Diagnostics.Monit
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Diagnostics.Monitoring.UnitTests", "src\Tests\Microsoft.Diagnostics.Monitoring.UnitTests\Microsoft.Diagnostics.Monitoring.UnitTests.csproj", "{04427BC9-8F54-4633-9BBB-6661715BD271}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Diagnostics.Monitoring.UnitTests.ApiClient", "src\Tests\Microsoft.Diagnostics.Monitoring.UnitTests.ApiClient\Microsoft.Diagnostics.Monitoring.UnitTests.ApiClient.csproj", "{1C6CFABB-6514-4F57-8479-12C86B89BE81}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Checked|Any CPU = Checked|Any CPU
@@ -293,6 +295,46 @@ Global
 		{04427BC9-8F54-4633-9BBB-6661715BD271}.RelWithDebInfo|x64.Build.0 = Release|Any CPU
 		{04427BC9-8F54-4633-9BBB-6661715BD271}.RelWithDebInfo|x86.ActiveCfg = Release|Any CPU
 		{04427BC9-8F54-4633-9BBB-6661715BD271}.RelWithDebInfo|x86.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|Any CPU.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|Any CPU.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|ARM.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|ARM.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|ARM64.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|ARM64.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|x64.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|x64.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|x86.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Checked|x86.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|ARM.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|x64.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Debug|x86.Build.0 = Debug|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|ARM.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|ARM.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|ARM64.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|x64.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|x64.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|x86.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.Release|x86.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|Any CPU.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|Any CPU.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|ARM.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|ARM.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|ARM64.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|ARM64.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|x64.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|x64.Build.0 = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|x86.ActiveCfg = Release|Any CPU
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81}.RelWithDebInfo|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -307,6 +349,7 @@ Global
 		{5AB5A4B4-3A13-4D42-B42C-B9193E60F426} = {C7568468-1C79-4944-8136-18812A7F9EA7}
 		{FCEC25E0-9EE1-416E-A7C4-2898E5640157} = {C7568468-1C79-4944-8136-18812A7F9EA7}
 		{04427BC9-8F54-4633-9BBB-6661715BD271} = {C7568468-1C79-4944-8136-18812A7F9EA7}
+		{1C6CFABB-6514-4F57-8479-12C86B89BE81} = {C7568468-1C79-4944-8136-18812A7F9EA7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {46465737-C938-44FC-BE1A-4CE139EBB5E0}
