--- src/Tests/Microsoft.Diagnostics.Monitoring.UnitTests/HttpApi/ContentTypes.cs
@@ -1,13 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information
-
-namespace Microsoft.Diagnostics.Monitoring.UnitTests.HttpApi
-{
-    internal static class ContentTypes
-    {
-        public const string ApplicationJson = "application/json";
-        public const string ApplicationProblemJson = "application/problem+json";
-        public const string TextPlain = "text/plain";
-    }
-}
